+++ elasticsearch/helpers/vectorstore/__init__.py
@@ -42,21 +42,21 @@ from elasticsearch.helpers.vectorstore._sync.vectorstore import VectorStore
 from elasticsearch.helpers.vectorstore._utils import DistanceMetric
 
 __all__ = [
-    "BM25Strategy",
-    "DenseVectorStrategy",
-    "DenseVectorScriptScoreStrategy",
-    "ElasticsearchEmbeddings",
-    "EmbeddingService",
-    "RetrievalStrategy",
-    "SparseVectorStrategy",
-    "VectorStore",
     "AsyncBM25Strategy",
-    "AsyncDenseVectorStrategy",
     "AsyncDenseVectorScriptScoreStrategy",
+    "AsyncDenseVectorStrategy",
     "AsyncElasticsearchEmbeddings",
     "AsyncEmbeddingService",
     "AsyncRetrievalStrategy",
     "AsyncSparseVectorStrategy",
     "AsyncVectorStore",
+    "BM25Strategy",
+    "DenseVectorScriptScoreStrategy",
+    "DenseVectorStrategy",
     "DistanceMetric",
+    "ElasticsearchEmbeddings",
+    "EmbeddingService",
+    "RetrievalStrategy",
+    "SparseVectorStrategy",
+    "VectorStore",
 ]
