+++ elasticsearch/helpers/vectorstore/_sync/vectorstore.py
@@ -22,8 +22,7 @@ from typing import Any, Callable, Dict, List, Optional
 from elasticsearch import Elasticsearch
 from elasticsearch._version import __versionstr__ as lib_version
 from elasticsearch.helpers import BulkIndexError, bulk
-from elasticsearch.helpers.vectorstore._sync.embedding_service import EmbeddingService
-from elasticsearch.helpers.vectorstore._sync.strategies import RetrievalStrategy
+from elasticsearch.helpers.vectorstore import EmbeddingService, RetrievalStrategy
 from elasticsearch.helpers.vectorstore._utils import maximal_marginal_relevance
 
 logger = logging.getLogger(__name__)
