+++ test_elasticsearch/test_server/test_helpers_vectorstore/__init__.py
@@ -14,3 +14,68 @@
 #  KIND, either express or implied.  See the License for the
 #  specific language governing permissions and limitations
 #  under the License.
+
+from typing import List
+
+from elastic_transport import Transport
+
+from elasticsearch.helpers.vectorstore import EmbeddingService
+
+
+class RequestSavingTransport(Transport):
+    def __init__(self, *args, **kwargs) -> None:
+        super().__init__(*args, **kwargs)
+        self.requests: list = []
+
+    def perform_request(self, *args, **kwargs):
+        self.requests.append(kwargs)
+        return super().perform_request(*args, **kwargs)
+
+
+class FakeEmbeddings(EmbeddingService):
+    """Fake embeddings functionality for testing."""
+
+    def __init__(self, dimensionality: int = 10) -> None:
+        self.dimensionality = dimensionality
+
+    def embed_documents(self, texts: List[str]) -> List[List[float]]:
+        """Return simple embeddings. Embeddings encode each text as its index."""
+        return [
+            [float(1.0)] * (self.dimensionality - 1) + [float(i)]
+            for i in range(len(texts))
+        ]
+
+    def embed_query(self, text: str) -> List[float]:
+        """Return constant query embeddings.
+        Embeddings are identical to embed_documents(texts)[0].
+        Distance to each text will be that text's index,
+        as it was passed to embed_documents.
+        """
+        return [float(1.0)] * (self.dimensionality - 1) + [float(0.0)]
+
+
+class ConsistentFakeEmbeddings(FakeEmbeddings):
+    """Fake embeddings which remember all the texts seen so far to return consistent
+    vectors for the same texts."""
+
+    def __init__(self, dimensionality: int = 10) -> None:
+        self.known_texts: List[str] = []
+        self.dimensionality = dimensionality
+
+    def embed_documents(self, texts: List[str]) -> List[List[float]]:
+        """Return consistent embeddings for each text seen so far."""
+        out_vectors = []
+        for text in texts:
+            if text not in self.known_texts:
+                self.known_texts.append(text)
+            vector = [float(1.0)] * (self.dimensionality - 1) + [
+                float(self.known_texts.index(text))
+            ]
+            out_vectors.append(vector)
+        return out_vectors
+
+    def embed_query(self, text: str) -> List[float]:
+        """Return consistent embeddings for the text, if seen before, or a constant
+        one if the text is unknown."""
+        result = self.embed_documents([text])
+        return result[0]
